requirements.txt
fastapi>=0.104.0
uvicorn>=0.24.0
pywebview>=4.4
pyyaml>=6.0
opencv-python>=4.8.0
numpy>=1.24.0
pytesseract>=0.3.10
pydantic>=2.5.0
aiosqlite>=0.19.0
